start reorg
