+++ zio-schema/shared/src/main/scala/zio/schema/diff/ProductDiffer.scala
@@ -1,0 +1,439 @@
+package zio.schema.diff
+
+
+import zio.schema._
+import scala.collection.immutable.ListMap
+import zio.Chunk
+
+//scalafmt: { maxColumn = 400, optIn.configStyleArguments = false }
+private[schema] object ProductDiffer {
+
+  def product0[Z](schema: Schema.CaseClass0[Z]): Differ[Z] = {
+    val _ = schema
+    (_: Z, _: Z) => Diff.identical
+  }
+
+  def product1[A, Z](schema: Schema.CaseClass1[A, Z]): Differ[Z] =
+    (thisZ: Z, thatZ: Z) =>
+      Diff
+        .Record[Z](ListMap(fieldDiffer(schema.field, schema.extractField)(thisZ, thatZ)), schema)
+        .orIdentical
+
+  def product2[A1, A2, Z](schema: Schema.CaseClass2[A1, A2, Z]): Differ[Z] =
+    (thisZ: Z, thatZ: Z) =>
+      Diff
+        .Record(ListMap.empty ++ Chunk(fieldDiffer(schema.field1, schema.extractField1), fieldDiffer(schema.field2, schema.extractField2)).map(_.apply(thisZ, thatZ)), schema)
+        .orIdentical
+
+  def product3[A1, A2, A3, Z](schema: Schema.CaseClass3[A1, A2, A3, Z]): Differ[Z] =
+    (thisZ: Z, thatZ: Z) =>
+      Diff
+        .Record(ListMap.empty ++ Chunk(fieldDiffer(schema.field1, schema.extractField1), fieldDiffer(schema.field2, schema.extractField2), fieldDiffer(schema.field3, schema.extractField3)).map(_.apply(thisZ, thatZ)), schema)
+        .orIdentical
+
+  def product4[A1, A2, A3, A4, Z](schema: Schema.CaseClass4[A1, A2, A3, A4, Z]): Differ[Z] =
+    (thisZ: Z, thatZ: Z) =>
+      Diff
+        .Record(ListMap.empty ++ Chunk(fieldDiffer(schema.field1, schema.extractField1), fieldDiffer(schema.field2, schema.extractField2), fieldDiffer(schema.field3, schema.extractField3), fieldDiffer(schema.field4, schema.extractField4)).map(_.apply(thisZ, thatZ)), schema)
+        .orIdentical
+
+  def product5[A1, A2, A3, A4, A5, Z](schema: Schema.CaseClass5[A1, A2, A3, A4, A5, Z]): Differ[Z] =
+    (thisZ: Z, thatZ: Z) =>
+      Diff
+        .Record(ListMap.empty ++ Chunk(fieldDiffer(schema.field1, schema.extractField1), fieldDiffer(schema.field2, schema.extractField2), fieldDiffer(schema.field3, schema.extractField3), fieldDiffer(schema.field4, schema.extractField4), fieldDiffer(schema.field5, schema.extractField5)).map(_.apply(thisZ, thatZ)), schema)
+        .orIdentical
+
+  def product6[A1, A2, A3, A4, A5, A6, Z](schema: Schema.CaseClass6[A1, A2, A3, A4, A5, A6, Z]): Differ[Z] =
+    (thisZ: Z, thatZ: Z) =>
+      Diff
+        .Record(ListMap.empty ++ Chunk(fieldDiffer(schema.field1, schema.extractField1), fieldDiffer(schema.field2, schema.extractField2), fieldDiffer(schema.field3, schema.extractField3), fieldDiffer(schema.field4, schema.extractField4), fieldDiffer(schema.field5, schema.extractField5), fieldDiffer(schema.field6, schema.extractField6)).map(_.apply(thisZ, thatZ)), schema)
+        .orIdentical
+
+  def product7[A1, A2, A3, A4, A5, A6, A7, Z](schema: Schema.CaseClass7[A1, A2, A3, A4, A5, A6, A7, Z]): Differ[Z] =
+    (thisZ: Z, thatZ: Z) =>
+      Diff
+        .Record(
+          ListMap.empty ++ Chunk(fieldDiffer(schema.field1, schema.extractField1), fieldDiffer(schema.field2, schema.extractField2), fieldDiffer(schema.field3, schema.extractField3), fieldDiffer(schema.field4, schema.extractField4), fieldDiffer(schema.field5, schema.extractField5), fieldDiffer(schema.field6, schema.extractField6), fieldDiffer(schema.field7, schema.extractField7))
+            .map(_.apply(thisZ, thatZ)),
+          schema
+        )
+        .orIdentical
+
+  def product8[A1, A2, A3, A4, A5, A6, A7, A8, Z](schema: Schema.CaseClass8[A1, A2, A3, A4, A5, A6, A7, A8, Z]): Differ[Z] =
+    (thisZ: Z, thatZ: Z) =>
+      Diff
+        .Record(
+          ListMap.empty ++ Chunk(
+            fieldDiffer(schema.field1, schema.extractField1),
+            fieldDiffer(schema.field2, schema.extractField2),
+            fieldDiffer(schema.field3, schema.extractField3),
+            fieldDiffer(schema.field4, schema.extractField4),
+            fieldDiffer(schema.field5, schema.extractField5),
+            fieldDiffer(schema.field6, schema.extractField6),
+            fieldDiffer(schema.field7, schema.extractField7),
+            fieldDiffer(schema.field8, schema.extractField8)
+          ).map(_.apply(thisZ, thatZ)),
+          schema
+        )
+        .orIdentical
+
+  def product9[A1, A2, A3, A4, A5, A6, A7, A8, A9, Z](schema: Schema.CaseClass9[A1, A2, A3, A4, A5, A6, A7, A8, A9, Z]): Differ[Z] =
+    (thisZ: Z, thatZ: Z) =>
+      Diff
+        .Record(
+          ListMap.empty ++ Chunk(
+            fieldDiffer(schema.field1, schema.extractField1),
+            fieldDiffer(schema.field2, schema.extractField2),
+            fieldDiffer(schema.field3, schema.extractField3),
+            fieldDiffer(schema.field4, schema.extractField4),
+            fieldDiffer(schema.field5, schema.extractField5),
+            fieldDiffer(schema.field6, schema.extractField6),
+            fieldDiffer(schema.field7, schema.extractField7),
+            fieldDiffer(schema.field8, schema.extractField8),
+            fieldDiffer(schema.field9, schema.extractField9)
+          ).map(_.apply(thisZ, thatZ)),
+          schema
+        )
+        .orIdentical
+
+  def product10[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, Z](schema: Schema.CaseClass10[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, Z]): Differ[Z] =
+    (thisZ: Z, thatZ: Z) =>
+      Diff
+        .Record(
+          ListMap.empty ++ Chunk(
+            fieldDiffer(schema.field1, schema.extractField1),
+            fieldDiffer(schema.field2, schema.extractField2),
+            fieldDiffer(schema.field3, schema.extractField3),
+            fieldDiffer(schema.field4, schema.extractField4),
+            fieldDiffer(schema.field5, schema.extractField5),
+            fieldDiffer(schema.field6, schema.extractField6),
+            fieldDiffer(schema.field7, schema.extractField7),
+            fieldDiffer(schema.field8, schema.extractField8),
+            fieldDiffer(schema.field9, schema.extractField9),
+            fieldDiffer(schema.field10, schema.extractField10)
+          ).map(_.apply(thisZ, thatZ)),
+          schema
+        )
+        .orIdentical
+
+  def product11[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, A11, Z](schema: Schema.CaseClass11[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, A11, Z]): Differ[Z] =
+    (thisZ: Z, thatZ: Z) =>
+      Diff
+        .Record(
+          ListMap.empty ++ Chunk(
+            fieldDiffer(schema.field1, schema.extractField1),
+            fieldDiffer(schema.field2, schema.extractField2),
+            fieldDiffer(schema.field3, schema.extractField3),
+            fieldDiffer(schema.field4, schema.extractField4),
+            fieldDiffer(schema.field5, schema.extractField5),
+            fieldDiffer(schema.field6, schema.extractField6),
+            fieldDiffer(schema.field7, schema.extractField7),
+            fieldDiffer(schema.field8, schema.extractField8),
+            fieldDiffer(schema.field9, schema.extractField9),
+            fieldDiffer(schema.field10, schema.extractField10),
+            fieldDiffer(schema.field11, schema.extractField11)
+          ).map(_.apply(thisZ, thatZ)),
+          schema
+        )
+        .orIdentical
+
+  def product12[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, A11, A12, Z](schema: Schema.CaseClass12[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, A11, A12, Z]): Differ[Z] =
+    (thisZ: Z, thatZ: Z) =>
+      Diff
+        .Record(
+          ListMap.empty ++ Chunk(
+            fieldDiffer(schema.field1, schema.extractField1),
+            fieldDiffer(schema.field2, schema.extractField2),
+            fieldDiffer(schema.field3, schema.extractField3),
+            fieldDiffer(schema.field4, schema.extractField4),
+            fieldDiffer(schema.field5, schema.extractField5),
+            fieldDiffer(schema.field6, schema.extractField6),
+            fieldDiffer(schema.field7, schema.extractField7),
+            fieldDiffer(schema.field8, schema.extractField8),
+            fieldDiffer(schema.field9, schema.extractField9),
+            fieldDiffer(schema.field10, schema.extractField10),
+            fieldDiffer(schema.field11, schema.extractField11),
+            fieldDiffer(schema.field12, schema.extractField12)
+          ).map(_.apply(thisZ, thatZ)),
+          schema
+        )
+        .orIdentical
+
+  def product13[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, A11, A12, A13, Z](schema: Schema.CaseClass13[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, A11, A12, A13, Z]): Differ[Z] =
+    (thisZ: Z, thatZ: Z) =>
+      Diff
+        .Record(
+          ListMap.empty ++ Chunk(
+            fieldDiffer(schema.field1, schema.extractField1),
+            fieldDiffer(schema.field2, schema.extractField2),
+            fieldDiffer(schema.field3, schema.extractField3),
+            fieldDiffer(schema.field4, schema.extractField4),
+            fieldDiffer(schema.field5, schema.extractField5),
+            fieldDiffer(schema.field6, schema.extractField6),
+            fieldDiffer(schema.field7, schema.extractField7),
+            fieldDiffer(schema.field8, schema.extractField8),
+            fieldDiffer(schema.field9, schema.extractField9),
+            fieldDiffer(schema.field10, schema.extractField10),
+            fieldDiffer(schema.field11, schema.extractField11),
+            fieldDiffer(schema.field12, schema.extractField12),
+            fieldDiffer(schema.field13, schema.extractField13)
+          ).map(_.apply(thisZ, thatZ)),
+          schema
+        )
+        .orIdentical
+
+  def product14[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, A11, A12, A13, A14, Z](schema: Schema.CaseClass14[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, A11, A12, A13, A14, Z]): Differ[Z] =
+    (thisZ: Z, thatZ: Z) =>
+      Diff
+        .Record(
+          ListMap.empty ++ Chunk(
+            fieldDiffer(schema.field1, schema.extractField1),
+            fieldDiffer(schema.field2, schema.extractField2),
+            fieldDiffer(schema.field3, schema.extractField3),
+            fieldDiffer(schema.field4, schema.extractField4),
+            fieldDiffer(schema.field5, schema.extractField5),
+            fieldDiffer(schema.field6, schema.extractField6),
+            fieldDiffer(schema.field7, schema.extractField7),
+            fieldDiffer(schema.field8, schema.extractField8),
+            fieldDiffer(schema.field9, schema.extractField9),
+            fieldDiffer(schema.field10, schema.extractField10),
+            fieldDiffer(schema.field11, schema.extractField11),
+            fieldDiffer(schema.field12, schema.extractField12),
+            fieldDiffer(schema.field13, schema.extractField13),
+            fieldDiffer(schema.field14, schema.extractField14)
+          ).map(_.apply(thisZ, thatZ)),
+          schema
+        )
+        .orIdentical
+
+  def product15[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, A11, A12, A13, A14, A15, Z](schema: Schema.CaseClass15[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, A11, A12, A13, A14, A15, Z]): Differ[Z] =
+    (thisZ: Z, thatZ: Z) =>
+      Diff
+        .Record(
+          ListMap.empty ++ Chunk(
+            fieldDiffer(schema.field1, schema.extractField1),
+            fieldDiffer(schema.field2, schema.extractField2),
+            fieldDiffer(schema.field3, schema.extractField3),
+            fieldDiffer(schema.field4, schema.extractField4),
+            fieldDiffer(schema.field5, schema.extractField5),
+            fieldDiffer(schema.field6, schema.extractField6),
+            fieldDiffer(schema.field7, schema.extractField7),
+            fieldDiffer(schema.field8, schema.extractField8),
+            fieldDiffer(schema.field9, schema.extractField9),
+            fieldDiffer(schema.field10, schema.extractField10),
+            fieldDiffer(schema.field11, schema.extractField11),
+            fieldDiffer(schema.field12, schema.extractField12),
+            fieldDiffer(schema.field13, schema.extractField13),
+            fieldDiffer(schema.field14, schema.extractField14),
+            fieldDiffer(schema.field15, schema.extractField15)
+          ).map(_.apply(thisZ, thatZ)),
+          schema
+        )
+        .orIdentical
+
+  def product16[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, A11, A12, A13, A14, A15, A16, Z](schema: Schema.CaseClass16[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, A11, A12, A13, A14, A15, A16, Z]): Differ[Z] =
+    (thisZ: Z, thatZ: Z) =>
+      Diff
+        .Record(
+          ListMap.empty ++ Chunk(
+            fieldDiffer(schema.field1, schema.extractField1),
+            fieldDiffer(schema.field2, schema.extractField2),
+            fieldDiffer(schema.field3, schema.extractField3),
+            fieldDiffer(schema.field4, schema.extractField4),
+            fieldDiffer(schema.field5, schema.extractField5),
+            fieldDiffer(schema.field6, schema.extractField6),
+            fieldDiffer(schema.field7, schema.extractField7),
+            fieldDiffer(schema.field8, schema.extractField8),
+            fieldDiffer(schema.field9, schema.extractField9),
+            fieldDiffer(schema.field10, schema.extractField10),
+            fieldDiffer(schema.field11, schema.extractField11),
+            fieldDiffer(schema.field12, schema.extractField12),
+            fieldDiffer(schema.field13, schema.extractField13),
+            fieldDiffer(schema.field14, schema.extractField14),
+            fieldDiffer(schema.field15, schema.extractField15),
+            fieldDiffer(schema.field16, schema.extractField16)
+          ).map(_.apply(thisZ, thatZ)),
+          schema
+        )
+        .orIdentical
+
+  def product17[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, A11, A12, A13, A14, A15, A16, A17, Z](schema: Schema.CaseClass17[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, A11, A12, A13, A14, A15, A16, A17, Z]): Differ[Z] =
+    (thisZ: Z, thatZ: Z) =>
+      Diff
+        .Record(
+          ListMap.empty ++ Chunk(
+            fieldDiffer(schema.field1, schema.extractField1),
+            fieldDiffer(schema.field2, schema.extractField2),
+            fieldDiffer(schema.field3, schema.extractField3),
+            fieldDiffer(schema.field4, schema.extractField4),
+            fieldDiffer(schema.field5, schema.extractField5),
+            fieldDiffer(schema.field6, schema.extractField6),
+            fieldDiffer(schema.field7, schema.extractField7),
+            fieldDiffer(schema.field8, schema.extractField8),
+            fieldDiffer(schema.field9, schema.extractField9),
+            fieldDiffer(schema.field10, schema.extractField10),
+            fieldDiffer(schema.field11, schema.extractField11),
+            fieldDiffer(schema.field12, schema.extractField12),
+            fieldDiffer(schema.field13, schema.extractField13),
+            fieldDiffer(schema.field14, schema.extractField14),
+            fieldDiffer(schema.field15, schema.extractField15),
+            fieldDiffer(schema.field16, schema.extractField16),
+            fieldDiffer(schema.field17, schema.extractField17)
+          ).map(_.apply(thisZ, thatZ)),
+          schema
+        )
+        .orIdentical
+
+  def product18[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, A11, A12, A13, A14, A15, A16, A17, A18, Z](schema: Schema.CaseClass18[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, A11, A12, A13, A14, A15, A16, A17, A18, Z]): Differ[Z] =
+    (thisZ: Z, thatZ: Z) =>
+      Diff
+        .Record(
+          ListMap.empty ++ Chunk(
+            fieldDiffer(schema.field1, schema.extractField1),
+            fieldDiffer(schema.field2, schema.extractField2),
+            fieldDiffer(schema.field3, schema.extractField3),
+            fieldDiffer(schema.field4, schema.extractField4),
+            fieldDiffer(schema.field5, schema.extractField5),
+            fieldDiffer(schema.field6, schema.extractField6),
+            fieldDiffer(schema.field7, schema.extractField7),
+            fieldDiffer(schema.field8, schema.extractField8),
+            fieldDiffer(schema.field9, schema.extractField9),
+            fieldDiffer(schema.field10, schema.extractField10),
+            fieldDiffer(schema.field11, schema.extractField11),
+            fieldDiffer(schema.field12, schema.extractField12),
+            fieldDiffer(schema.field13, schema.extractField13),
+            fieldDiffer(schema.field14, schema.extractField14),
+            fieldDiffer(schema.field15, schema.extractField15),
+            fieldDiffer(schema.field16, schema.extractField16),
+            fieldDiffer(schema.field17, schema.extractField17),
+            fieldDiffer(schema.field18, schema.extractField18)
+          ).map(_.apply(thisZ, thatZ)),
+          schema
+        )
+        .orIdentical
+
+  def product19[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, A11, A12, A13, A14, A15, A16, A17, A18, A19, Z](schema: Schema.CaseClass19[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, A11, A12, A13, A14, A15, A16, A17, A18, A19, Z]): Differ[Z] =
+    (thisZ: Z, thatZ: Z) =>
+      Diff
+        .Record(
+          ListMap.empty ++ Chunk(
+            fieldDiffer(schema.field1, schema.extractField1),
+            fieldDiffer(schema.field2, schema.extractField2),
+            fieldDiffer(schema.field3, schema.extractField3),
+            fieldDiffer(schema.field4, schema.extractField4),
+            fieldDiffer(schema.field5, schema.extractField5),
+            fieldDiffer(schema.field6, schema.extractField6),
+            fieldDiffer(schema.field7, schema.extractField7),
+            fieldDiffer(schema.field8, schema.extractField8),
+            fieldDiffer(schema.field9, schema.extractField9),
+            fieldDiffer(schema.field10, schema.extractField10),
+            fieldDiffer(schema.field11, schema.extractField11),
+            fieldDiffer(schema.field12, schema.extractField12),
+            fieldDiffer(schema.field13, schema.extractField13),
+            fieldDiffer(schema.field14, schema.extractField14),
+            fieldDiffer(schema.field15, schema.extractField15),
+            fieldDiffer(schema.field16, schema.extractField16),
+            fieldDiffer(schema.field17, schema.extractField17),
+            fieldDiffer(schema.field18, schema.extractField18),
+            fieldDiffer(schema.field19, schema.extractField19)
+          ).map(_.apply(thisZ, thatZ)),
+          schema
+        )
+        .orIdentical
+
+  def product20[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, A11, A12, A13, A14, A15, A16, A17, A18, A19, A20, Z](schema: Schema.CaseClass20[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, A11, A12, A13, A14, A15, A16, A17, A18, A19, A20, Z]): Differ[Z] =
+    (thisZ: Z, thatZ: Z) =>
+      Diff
+        .Record(
+          ListMap.empty ++ Chunk(
+            fieldDiffer(schema.field1, schema.extractField1),
+            fieldDiffer(schema.field2, schema.extractField2),
+            fieldDiffer(schema.field3, schema.extractField3),
+            fieldDiffer(schema.field4, schema.extractField4),
+            fieldDiffer(schema.field5, schema.extractField5),
+            fieldDiffer(schema.field6, schema.extractField6),
+            fieldDiffer(schema.field7, schema.extractField7),
+            fieldDiffer(schema.field8, schema.extractField8),
+            fieldDiffer(schema.field9, schema.extractField9),
+            fieldDiffer(schema.field10, schema.extractField10),
+            fieldDiffer(schema.field11, schema.extractField11),
+            fieldDiffer(schema.field12, schema.extractField12),
+            fieldDiffer(schema.field13, schema.extractField13),
+            fieldDiffer(schema.field14, schema.extractField14),
+            fieldDiffer(schema.field15, schema.extractField15),
+            fieldDiffer(schema.field16, schema.extractField16),
+            fieldDiffer(schema.field17, schema.extractField17),
+            fieldDiffer(schema.field18, schema.extractField18),
+            fieldDiffer(schema.field19, schema.extractField19),
+            fieldDiffer(schema.field20, schema.extractField20)
+          ).map(_.apply(thisZ, thatZ)),
+          schema
+        )
+        .orIdentical
+
+  def product21[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, A11, A12, A13, A14, A15, A16, A17, A18, A19, A20, A21, Z](schema: Schema.CaseClass21[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, A11, A12, A13, A14, A15, A16, A17, A18, A19, A20, A21, Z]): Differ[Z] =
+    (thisZ: Z, thatZ: Z) =>
+      Diff
+        .Record(
+          ListMap.empty ++ Chunk(
+            fieldDiffer(schema.field1, schema.extractField1),
+            fieldDiffer(schema.field2, schema.extractField2),
+            fieldDiffer(schema.field3, schema.extractField3),
+            fieldDiffer(schema.field4, schema.extractField4),
+            fieldDiffer(schema.field5, schema.extractField5),
+            fieldDiffer(schema.field6, schema.extractField6),
+            fieldDiffer(schema.field7, schema.extractField7),
+            fieldDiffer(schema.field8, schema.extractField8),
+            fieldDiffer(schema.field9, schema.extractField9),
+            fieldDiffer(schema.field10, schema.extractField10),
+            fieldDiffer(schema.field11, schema.extractField11),
+            fieldDiffer(schema.field12, schema.extractField12),
+            fieldDiffer(schema.field13, schema.extractField13),
+            fieldDiffer(schema.field14, schema.extractField14),
+            fieldDiffer(schema.field15, schema.extractField15),
+            fieldDiffer(schema.field16, schema.extractField16),
+            fieldDiffer(schema.field17, schema.extractField17),
+            fieldDiffer(schema.field18, schema.extractField18),
+            fieldDiffer(schema.field19, schema.extractField19),
+            fieldDiffer(schema.field20, schema.extractField20),
+            fieldDiffer(schema.field21, schema.extractField21)
+          ).map(_.apply(thisZ, thatZ)),
+          schema
+        )
+        .orIdentical
+
+  def product22[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, A11, A12, A13, A14, A15, A16, A17, A18, A19, A20, A21, A22, Z](schema: Schema.CaseClass22[A1, A2, A3, A4, A5, A6, A7, A8, A9, A10, A11, A12, A13, A14, A15, A16, A17, A18, A19, A20, A21, A22, Z]): Differ[Z] =
+    (thisZ: Z, thatZ: Z) =>
+      Diff
+        .Record(
+          ListMap.empty ++ Chunk(
+            fieldDiffer(schema.field1, schema.extractField1),
+            fieldDiffer(schema.field2, schema.extractField2),
+            fieldDiffer(schema.field3, schema.extractField3),
+            fieldDiffer(schema.field4, schema.extractField4),
+            fieldDiffer(schema.field5, schema.extractField5),
+            fieldDiffer(schema.field6, schema.extractField6),
+            fieldDiffer(schema.field7, schema.extractField7),
+            fieldDiffer(schema.field8, schema.extractField8),
+            fieldDiffer(schema.field9, schema.extractField9),
+            fieldDiffer(schema.field10, schema.extractField10),
+            fieldDiffer(schema.field11, schema.extractField11),
+            fieldDiffer(schema.field12, schema.extractField12),
+            fieldDiffer(schema.field13, schema.extractField13),
+            fieldDiffer(schema.field14, schema.extractField14),
+            fieldDiffer(schema.field15, schema.extractField15),
+            fieldDiffer(schema.field16, schema.extractField16),
+            fieldDiffer(schema.field17, schema.extractField17),
+            fieldDiffer(schema.field18, schema.extractField18),
+            fieldDiffer(schema.field19, schema.extractField19),
+            fieldDiffer(schema.field20, schema.extractField20),
+            fieldDiffer(schema.field21, schema.extractField21),
+            fieldDiffer(schema.field22, schema.extractField22)
+          ).map(_.apply(thisZ, thatZ)),
+          schema
+        )
+        .orIdentical
+
+  private def fieldDiffer[A, Z](field: Schema.Field[A], extract: Z => A): (Z, Z) => (String, Diff[A]) =
+    (thisZ: Z, thatZ: Z) => field.label -> Differ.fromSchema(field.schema)(extract(thisZ), extract(thatZ))
+}
